graph.pdf is the clamp example
It works, but we might want to work on error handling next.
The decompiler becomes untrustworthy if it outputs incomplete
code for nodes that it doesn't know how to handle.

I've ran into this when running it over the matrix multiply
example -- we don't know how to handle loops yet.