WIP, started looking into loops
Tobi dropped by the pod to chat with the Storefront Renderer team and
the conversations were so interesting that I couldn't get much done.

It's definitely not because I only had four and a half hours of sleep
last night.

But, I actually feel pretty good about what I got done today. Maybe it's
the lack of sleep talking though.