We output somewhat sensible code for all examples
Huh, much of the Ruby graphs are made of subgraphs that do the method
lookup logic. Once I got that to work everything kind of snapped into
place.

loops and phi are interesting. I suppose LoopBeginNode.is_a?(MergeNode)
and its ends are LoopEndNodes.

Much of the code is crap, but we can go clean it up later if we want
to go deeper in this direction.